[chore] prometheusreceiver: move helper funcs and tests to utils, remove duplicate test code
